move everything into snippet/browser
