+++ src/vs/editor/browser/editor.all.ts
@@ -38,8 +38,7 @@ import 'vs/editor/contrib/quickFix/browser/quickFixCommands';
 import 'vs/editor/contrib/referenceSearch/browser/referenceSearch';
 import 'vs/editor/contrib/rename/browser/rename';
 import 'vs/editor/contrib/smartSelect/common/smartSelect';
-import 'vs/editor/contrib/snippet/common/snippet';
-import 'vs/editor/contrib/snippet/browser/snippet';
+import 'vs/editor/contrib/snippet/browser/snippetController2';
 import 'vs/editor/contrib/suggest/browser/suggestController';
 import 'vs/editor/contrib/toggleTabFocusMode/common/toggleTabFocusMode';
 import 'vs/editor/contrib/wordHighlighter/common/wordHighlighter';
