+++ src/vs/editor/contrib/snippet/browser/snippetParser.ts
