+++ src/vs/editor/contrib/snippet/browser/snippetSession.css
