+++ src/vs/editor/contrib/snippet/browser/snippetSession.ts
@@ -5,16 +5,17 @@
 
 'use strict';
 
+import 'vs/css!./snippetSession';
 import { getLeadingWhitespace } from 'vs/base/common/strings';
 import { ICommonCodeEditor, IModel, TrackedRangeStickiness, IIdentifiedSingleEditOperation } from 'vs/editor/common/editorCommon';
 import { EditOperation } from 'vs/editor/common/core/editOperation';
-import { TextmateSnippet, Placeholder, SnippetParser } from '../common/snippetParser';
+import { TextmateSnippet, Placeholder, SnippetParser } from './snippetParser';
 import { Selection } from 'vs/editor/common/core/selection';
 import { Range } from 'vs/editor/common/core/range';
 import { IPosition } from 'vs/editor/common/core/position';
 import { groupBy } from 'vs/base/common/arrays';
 import { dispose } from 'vs/base/common/lifecycle';
-import { EditorSnippetVariableResolver } from "vs/editor/contrib/snippet/common/snippetVariables";
+import { EditorSnippetVariableResolver } from "./snippetVariables";
 import { ModelDecorationOptions } from "vs/editor/common/model/textModelWithDecorations";
 
 export class OneSnippet {
