+++ src/vs/editor/contrib/snippet/browser/snippetVariables.ts
