+++ src/vs/workbench/parts/snippets/electron-browser/TMSnippets.ts
@@ -15,8 +15,8 @@ import { ISnippetsService, ISnippet } from 'vs/workbench/parts/snippets/electron
 import { IModeService } from 'vs/editor/common/services/modeService';
 import { languagesExtPoint } from 'vs/editor/common/services/modeServiceImpl';
 import { LanguageIdentifier } from 'vs/editor/common/modes';
-import { SnippetParser, Marker, Placeholder, Variable, Text, walk } from 'vs/editor/contrib/snippet/common/snippetParser';
-import { EditorSnippetVariableResolver } from "vs/editor/contrib/snippet/common/snippetVariables";
+import { SnippetParser, Marker, Placeholder, Variable, Text, walk } from 'vs/editor/contrib/snippet/browser/snippetParser';
+import { EditorSnippetVariableResolver } from "vs/editor/contrib/snippet/browser/snippetVariables";
 
 interface ISnippetsExtensionPoint {
 	language: string;
